+++ docs/.vitepress/theme/components/landing/2. feature-section/FeatureInstantServerStart.vue
@@ -344,7 +344,7 @@ onUnmounted(() => {
       height: 100%;
       border-radius: 12px 0 0 12px;
       background:
-        url('/noise.png'),
+        url('../common/noise.webp'),
         radial-gradient(
           ellipse 140% 80% at 96% bottom,
           #13b351 0%,
@@ -381,7 +381,7 @@ onUnmounted(() => {
     right: 40px;
     width: 1px;
     height: calc(100% - 170px - 33px);
-    background: url('/noise.png'), #13b351;
+    background: url('../common/noise.webp'), #13b351;
     box-shadow: 0 0 10px 0 #13b351;
     transition: all 0.5s ease-in;
     will-change: transform, opacity;