+++ packages/create-vite/template-svelte-ts/svelte.config.js
@@ -1,5 +1,6 @@
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
 
+/** @type {import("@sveltejs/vite-plugin-svelte").SvelteConfig} */
 export default {
   // Consult https://svelte.dev/docs#compile-time-svelte-preprocess
   // for more information about preprocessors